Refactor stuff out of the plug
I guess I need to wrap the graphql-erlang module itself in Elixir
modules so I can specify structs to make it nicer to work with. The
actual plug will use this work and at that point to plug will become
very small.

This has the added benefit that the GraphQL modules can be used in
other contexts than just web pipelines.